update to separate msg.rs and state.rs
